+++ src/back/controllers/api/routes/challenges/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const router = require('express').Router;
-const challengesController = require('./controller');
+const challengesController = require('../../../challenges/controller');
 
 module.exports = router()
     .get('/', challengesController.getAll)
