+++ src/front/app/store/enhancers.js
@@ -4,8 +4,7 @@ import { routerMiddleware } from 'react-router-redux';
 import thunk from 'redux-thunk';
 
 export default compose(
-    applyMiddleware(thunk),
-    applyMiddleware(routerMiddleware(history)),
+    applyMiddleware(thunk, routerMiddleware(history)),
     (process.env.NODE_ENV === 'development' && global.devToolsExtension)
         ? global.devToolsExtension() : f => f
 );
